+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -22,11 +22,23 @@
                 type="post"
                 :class="$style['c-checkout-form']"
                 @submit.prevent="onFormSubmit">
+                <p
+                    :class="[
+                        $style['c-checkout-link']
+                    ]">
+                    <a
+                        :href="loginUrl"
+                        data-test-id="switch-user-link"
+                        @click="onVisitLoginPage">
+                        {{ $t('switchUserText', { name }) }}
+                    </a>
+                </p>
                 <form-field
-                    v-model="customer.mobileNumber"
+                    :value="customer.mobileNumber"
                     name="mobile-number"
                     :label-text="$t('labels.mobileNumber')"
-                    :has-error="!isMobileNumberValid">
+                    :has-error="!isMobileNumberValid"
+                    @input="updateMobileNumber">
                     <template #error>
                         <error-message
                             v-if="!isMobileNumberValid"
@@ -119,6 +131,11 @@ export default {
             required: true
         },
 
+        createGuestUrl: {
+            type: String,
+            required: true
+        },
+
         checkoutTimeout: {
             type: Number,
             required: false,
@@ -131,9 +148,19 @@ export default {
             default: 1000
         },
 
+        createGuestTimeout: {
+            type: Number,
+            default: 1000
+        },
+
         authToken: {
             type: String,
             default: ''
+        },
+
+        loginUrl: {
+            type: String,
+            required: true
         }
     },
 
@@ -162,14 +189,15 @@ export default {
 
     computed: {
         ...mapState('checkout', [
-            'id',
-            'serviceType',
             'customer',
             'fulfilment',
-            'notes',
+            'id',
             'isFulfillable',
+            'isLoggedIn',
+            'messages',
+            'notes',
             'notices',
-            'messages'
+            'serviceType'
         ]),
 
         name () {
@@ -228,11 +256,17 @@ export default {
 
     methods: {
         ...mapActions('checkout', [
+            'createGuestUser',
+            'getAvailableFulfilment',
             'getCheckout',
             'postCheckout',
-            'getAvailableFulfilment',
-            'setAuthToken'
+            'setAuthToken',
+            'updateMobileNumber'
         ]),
+
+        onVisitLoginPage () {
+            this.$emit(EventNames.CheckoutVisitLoginPage);
+        },
 
         /**
          * Submit the checkout details while emitting events to communicate its success or failure.
@@ -248,6 +282,10 @@ export default {
                     checkoutData.address = this.fulfilment.address;
                 }
 
+                if (!this.isLoggedIn) {
+                    await this.setupGuestUser();
+                }
+
                 await this.postCheckout({
                     url: 'myPostUrl',
                     tenant: this.tenant,
@@ -259,6 +297,26 @@ export default {
             } catch (thrownErrors) {
                 this.$emit(EventNames.CheckoutFailure, thrownErrors);
             }
+        },
+
+        /**
+         * Setup a new guest user account. This method will be called when isLoggedIn is false.
+         *
+         */
+        async setupGuestUser () {
+            const createGuestData = {
+                emailAddress: this.customer.email,
+                firstName: this.customer.firstName,
+                lastName: this.customer.lastName,
+                registrationSource: 'Guest'
+            };
+
+            await this.createGuestUser({
+                url: this.createGuestUrl,
+                tenant: this.tenant,
+                data: createGuestData,
+                timeout: this.createGuestTimeout
+            });
         },
 
         /**
@@ -430,5 +488,19 @@ $checkout-padding                         : spacing(x5) 100px;
     .c-checkout-submitButton {
         margin: spacing(x4) 0 spacing(x0.5);
     }
+}
+
+.c-checkout-link {
+  text-align: center;
+
+  a {
+    text-decoration: none;
+    font-weight: $font-weight-bold;
+
+    &:hover,
+    &:focus {
+      text-decoration: underline;
+    }
+  }
 }
 </style>