+++ packages/components/organisms/f-checkout/src/store/checkout.module.js
@@ -8,6 +8,8 @@ export default {
         serviceType: '',
         customer: {
             firstName: '',
+            lastName: '',
+            email: '',
             mobileNumber: ''
         },
         fulfilment: {
@@ -54,7 +56,7 @@ export default {
 
             const { data } = await axios.get(url, config);
 
-            commit('updateState', data);
+            commit('UPDATE_STATE', data);
         },
 
         /**
@@ -88,6 +90,32 @@ export default {
         },
 
         /**
+         * Post the guest user details to the backend.
+         *
+         * @param {Object} commit - Automatically handled by Vuex to be able to commit mutations.
+         * @param {Object} state - Automatically handled by Vuex to be able to retrieve state.
+         * @param {Object} payload - Parameter with the different configurations for the request.
+         */
+        // eslint-disable-next-line no-unused-vars
+        createGuestUser: async ({ commit, state }, {
+            url, tenant, data, timeout
+        }) => {
+            const config = {
+                method: 'post',
+                headers: {
+                    'Content-Type': 'application/json',
+                    'Accept-Tenant': tenant
+                },
+                timeout
+            };
+
+            const response = await axios.post(url, data, config);
+            // eslint-disable-next-line no-unused-vars
+            const otac = response.data.token;
+            // TODO: Use otac to log the user in
+        },
+
+        /**
          * Get the fulfilment details from the backend and update the state.
          *
          * @param {Object} commit - Automatically handled by Vuex to be able to commit mutations.
@@ -106,16 +134,24 @@ export default {
 
             const { data } = await axios.get(url, config);
 
-            commit('updateAvailableFulfilment', data);
+            commit('UPDATE_AVAILABLE_FULFILMENT_TIMES', data);
         },
 
         setAuthToken: ({ commit }, authToken) => {
-            commit('updateAuth', authToken);
+            commit('UPDATE_AUTH', authToken);
+        },
+
+        updateFulfilmentAddress ({ commit }, payload) {
+            commit('UPDATE_FULFILMENT_ADDRESS', payload);
+        },
+
+        updateMobileNumber ({ commit }, payload) {
+            commit('UPDATE_MOBILE_NUMBER', payload);
         }
     },
 
     mutations: {
-        updateState: (state, {
+        UPDATE_STATE: (state, {
             id,
             serviceType,
             customer,
@@ -151,7 +187,7 @@ export default {
             state.messages = messages;
         },
 
-        updateAvailableFulfilment: (state, {
+        UPDATE_AVAILABLE_FULFILMENT_TIMES: (state, {
             times,
             asapAvailable
         }) => {
@@ -159,9 +195,20 @@ export default {
             state.availableFulfilment.isAsapAvailable = asapAvailable;
         },
 
-        updateAuth: (state, authToken) => {
+        UPDATE_AUTH: (state, authToken) => {
             state.authToken = authToken;
             state.isLoggedIn = !!authToken;
+        },
+
+        UPDATE_FULFILMENT_ADDRESS (state, address) {
+            state.fulfilment.address = {
+                ...state.fulfilment.address,
+                ...address
+            };
+        },
+
+        UPDATE_MOBILE_NUMBER (state, mobileNumber) {
+            state.customer.mobileNumber = mobileNumber;
         }
     }
 };