+++ packages/components/organisms/f-checkout/stories/Checkout.stories.js
@@ -6,7 +6,7 @@ import { withA11y } from '@storybook/addon-a11y';
 import {
     VALID_LOCALES,
     ENGLISH_LOCALE
-} from '../../../../storybook/constants/globalisation';
+} from '@justeat/storybook/constants/globalisation';
 
 import VueCheckout from '../src/components/Checkout.vue';
 import CheckoutMock from '../src/demo/checkoutMock';
@@ -16,10 +16,13 @@ Vue.use(Vuex);
 const deliveryUrl = '/checkout-delivery.json';
 const collectionUrl = '/checkout-collection.json';
 const checkoutAvailableFulfilmentUrl = '/checkout-available-fulfilment.json';
+const createGuestUrl = '/create-guest.json';
 
 CheckoutMock.setupCheckoutMethod(deliveryUrl);
 CheckoutMock.setupCheckoutMethod(collectionUrl);
 CheckoutMock.setupCheckoutMethod(checkoutAvailableFulfilmentUrl);
+CheckoutMock.setupCheckoutMethod(createGuestUrl);
+CheckoutMock.passThroughAny();
 
 export default {
     title: 'Components/Organisms',
@@ -38,14 +41,26 @@ export const CheckoutComponent = () => ({
         checkoutAvailableFulfilmentUrl: {
             default: select('Available Fulfilment Url', [checkoutAvailableFulfilmentUrl], checkoutAvailableFulfilmentUrl)
         },
+        createGuestUrl: {
+            default: text('Create Guest Url', createGuestUrl)
+        },
         authToken: {
             default: text('Auth token', '')
+        },
+        loginUrl: {
+            default: text('Login Url', '/login')
         }
     },
     store: new Vuex.Store({}),
-    template: '<vue-checkout'
-        + ' :checkoutUrl="checkoutUrl" :checkout-available-fulfilment-url="checkoutAvailableFulfilmentUrl" :authToken="authToken" :locale="locale"'
-        + ' :key="`${locale},${checkoutUrl},${checkoutAvailableFulfilmentUrl},${authToken}`" />'
+    template: '<vue-checkout ' +
+        ':checkoutUrl="checkoutUrl" ' +
+        ':checkout-available-fulfilment-url="checkoutAvailableFulfilmentUrl" ' +
+        ':create-guest-url="createGuestUrl" ' +
+        ':authToken="authToken" ' +
+        ':locale="locale" ' +
+        ':loginUrl="loginUrl" ' +
+        // eslint-disable-next-line no-template-curly-in-string
+        ' :key="`${locale},${checkoutUrl},${checkoutAvailableFulfilmentUrl},${authToken},${createGuestUrl}`" />'
 });
 
 CheckoutComponent.storyName = 'f-checkout';